requirements.txt
absl-py==0.11.0
astunparse==1.6.3
attrs==20.3.0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-WTF==0.14.3
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.7
Pillow==8.0.1
pluggy==0.13.1
protobuf==3.14.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.4
selenium==3.141.0
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3